+++ src/spellbot/services/channels.py
@@ -200,6 +200,18 @@ class ChannelsService:
         DatabaseSession.commit()
         return value
 
+    @sync_to_async()
+    def set_blind_games(self, xid: int, value: bool) -> bool:
+        query = (
+            update(Channel)  # type: ignore
+            .where(Channel.xid == xid)
+            .values(blind_games=value)
+            .execution_options(synchronize_session=False)
+        )
+        DatabaseSession.execute(query)
+        DatabaseSession.commit()
+        return value
+
     # TODO: Refactor how confirmation/points/ELO works.
     # @sync_to_async()
     # def set_require_confirmation(self, xid: int, value: bool) -> bool: